pyproject.toml
[tool.poetry]
name = "scarecrow-py"
version = "0.1.0"
description = ""
authors = ["Leonhard Horstmeyer <leo.h@hotmail.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-osc = "^1.8.3"
opencv-python = "^4.9.0.80"
google-cloud-vision = "^3.7.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
